move common analytics types to SDK
